Rename
   ffmpeg_conf.py       => ffmpeg_config.py
   ffmpeg_conf_defaults => ffmpeg_config_defaults.py

to better match the names in moviepy.
