+++ glumpy/ext/ffmpeg_config_defaults.py
